+++ resources/lang/ca/students.php
@@ -13,6 +13,7 @@ return [
     'excel' => 'Importa de l\'Excel',
     'excel_template' => 'Descarrega la plantilla',
     'dnd' => 'Arrossega el fitxer o clica per pujar',
+    'search' => 'Cerca per nom',
 
     'name_surname' => 'Nom i cognoms',
     'name' => 'Nom',