+++ resources/lang/en/students.php
@@ -13,6 +13,7 @@ return [
     'excel' => 'Import from Excel',
     'excel_template' => 'Download the template',
     'dnd' => 'Drop your file here or click to upload',
+    'search' => 'Search by name',
     
     'name_surname' => 'Name and surname',
     'name' => 'Name',