+++ resources/lang/es/students.php
@@ -13,6 +13,7 @@ return [
     'excel' => 'Importa del Excel',
     'excel_template' => 'Descarga la plantilla',
     'dnd' => 'Arrastra el fichero o haz clic para subir',
+    'search' => 'Busca por nombre',
     
     'name_surname' => 'Nombre y apellidos',
     'name' => 'Nombre',